Fix error with no content folder

+++ src/services/startupService.ts
@@ -5,8 +5,16 @@ import env from '../env';
 const startApplication = async () => {
   await configService.getConfig();
   await $`yt-dlp -U`;
-  await $`find ${env.CONTENT_FOLDER_PATH} -name "*.video" -type f -delete`;
-  await $`find ${env.CONTENT_FOLDER_PATH} -name "*.audio" -type f -delete`;
+
+  const contentFolderExists = await configService
+    .verifyContentFolderExists()
+    .then(() => true)
+    .catch(() => false);
+
+  if (contentFolderExists) {
+    await $`find ${env.CONTENT_FOLDER_PATH} -name "*.video" -type f -delete`;
+    await $`find ${env.CONTENT_FOLDER_PATH} -name "*.audio" -type f -delete`;
+  }
 };
 
 export default { startApplication };
